+++ LughSharp/Lugh/Graphics/GLUtils/ImmediateModeRenderer20.cs
@@ -110,15 +110,15 @@ public class ImmediateModeRenderer20 : IImmediateModeRenderer
         _vertices   = new float[ maxVertices * ( _mesh.VertexAttributes.VertexSize / 4 ) ];
         _vertexSize = _mesh.VertexAttributes.VertexSize / 4;
 
-        var attribute = _mesh.GetVertexAttribute( VertexAttributes.Usage.NORMAL );
+        var attribute = _mesh.GetVertexAttribute( ( int )VertexAttributes.Usage.NORMAL );
 
         _normalOffset = attribute != null ? attribute.Offset / 4 : 0;
 
-        attribute = _mesh.GetVertexAttribute( VertexAttributes.Usage.COLOR_PACKED );
+        attribute = _mesh.GetVertexAttribute( ( int )VertexAttributes.Usage.COLOR_PACKED );
 
         _colorOffset = attribute != null ? attribute.Offset / 4 : 0;
 
-        attribute = _mesh.GetVertexAttribute( VertexAttributes.Usage.TEXTURE_COORDINATES );
+        attribute = _mesh.GetVertexAttribute( ( int )VertexAttributes.Usage.TEXTURE_COORDINATES );
 
         _texCoordOffset = attribute != null ? attribute.Offset / 4 : 0;
 
@@ -275,26 +275,26 @@ public class ImmediateModeRenderer20 : IImmediateModeRenderer
     {
         var attribs = new List< VertexAttribute >
         {
-            new( VertexAttributes.Usage.POSITION, 3, ShaderProgram.POSITION_ATTRIBUTE ),
+            new( ( int )VertexAttributes.Usage.POSITION, 3, ShaderProgram.POSITION_ATTRIBUTE ),
         };
 
         if ( hasNormals )
         {
-            attribs.Add( new VertexAttribute( VertexAttributes.Usage.NORMAL,
+            attribs.Add( new VertexAttribute( ( int )VertexAttributes.Usage.NORMAL,
                                               3,
                                               ShaderProgram.NORMAL_ATTRIBUTE ) );
         }
 
         if ( hasColor )
         {
-            attribs.Add( new VertexAttribute( VertexAttributes.Usage.COLOR_PACKED,
+            attribs.Add( new VertexAttribute( ( int )VertexAttributes.Usage.COLOR_PACKED,
                                               4,
                                               ShaderProgram.COLOR_ATTRIBUTE ) );
         }
 
         for ( var i = 0; i < numTexCoords; i++ )
         {
-            attribs.Add( new VertexAttribute( VertexAttributes.Usage.TEXTURE_COORDINATES,
+            attribs.Add( new VertexAttribute( ( int )VertexAttributes.Usage.TEXTURE_COORDINATES,
                                               2,
                                               ShaderProgram.TEXCOORD_ATTRIBUTE + i ) );
         }
